requirements.txt
kaggle==1.5.3
Pillow==5.4.1
opencv-python==4.0.0.21
numpy==1.16.2
tensorflow-gpu==1.13.1
Click==7.0
